requirements_brainscore.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
anyio=3.7.1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
astor=0.8.1=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
boto3=1.33.13=pypi_0
botocore=1.33.13=pypi_0
brain-score=1.3=pypi_0
brainio=0.2.0=pypi_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
cftime=1.5.1.1=py37hce1f21e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.2.1=pypi_0
comm=0.1.4=pypi_0
cryptography=39.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
dask=2022.2.0=pypi_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=py37h06a4308_0
exceptiongroup=1.2.0=pypi_0
fastjsonschema=2.19.0=pypi_0
ffmpeg=4.3=hf484d3e_0
fire=0.5.0=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.1.0=pypi_0
gast=0.2.2=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-pasta=0.2.0=pypi_0
grpcio=1.60.0=pypi_0
h5py=3.8.0=pypi_0
hdf4=4.2.13=h3ca952b_2
hdf5=1.12.1=h70be1eb_2
idna=3.4=py37h06a4308_0
importlib-metadata=4.13.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.1=pypi_0
jedi=0.18.1=py37h06a4308_1
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyter_client=7.4.9=py37h06a4308_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h568e23c_1
lame=3.100=h7b6447c_0
latexcodec=2.0.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcurl=8.2.1=h91b91d3_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnetcdf=4.8.1=h424c3a0_4
libnghttp2=1.52.0=ha637b67_1
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libzip=1.8.0=h5cef20c_0
locket=1.0.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=py37h06a4308_0
mistune=3.0.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
model-template=0.1.0=pypi_0
model-tools=0.1.0=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.6.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py37h06a4308_0
netcdf4=1.6.2=py37h89d13dc_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
notebook=6.5.6=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=22.0=py37h06a4308_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.4.1=pypi_0
peewee=3.17.0=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=pypi_0
pip=22.3.1=py37h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.36=py37h06a4308_0
protobuf=3.20.3=pypi_0
psutil=5.9.0=py37h5eee18b_0
psycopg2-binary=2.9.9=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pybtex=0.24.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.0.0=py37h06a4308_0
pyparsing=3.1.1=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py37_1
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=23.2.0=py37h6a678d5_0
qtconsole=5.4.4=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=py37h06a4308_0
result-caching=0.1.0=pypi_0
s3transfer=0.8.2=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.2=pypi_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
tensorboard=1.15.0=pypi_0
tensorflow=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
toolz=0.12.0=pypi_0
torchvision=0.12.0=py37_cu113
tornado=6.2=py37h5eee18b_0
tqdm=4.66.1=pypi_0
traitlets=5.7.1=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
widgetsnbextension=4.0.9=pypi_0
wrapt=1.16.0=pypi_0
xarray=0.20.2=pypi_0
xz=5.4.5=h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
